Fix build with Microsoft Visual C++, Fix gfxhook and felica issue (#48)
I just wanna say that It is a SHAME that a Windows ONLY project was not able to build without MINGW
Also where's the missing `3mpxsc.h` in diva hook?

This also fixes the window size issue from hook_CreateWindowExA in gfxhook
And Fixes felica issue as described in #45

Reviewed-on: https://gitea.tendokyu.moe/Dniel97/segatools/pulls/48
Co-committed-by: GEEKiDoS <geek_ds@foxmail.com>

+++ board/led15093-cmd.h
@@ -88,7 +88,7 @@ struct led15093_req_reset {
 struct led15093_req_set_timeout {
     struct led15093_req_hdr hdr;
     uint8_t cmd;
-    uint8_t count;
+    uint16_t count;
 };
 
 struct led15093_req_set_disable_response {
