+++ idacio/xi.c
@@ -184,10 +184,10 @@ static int16_t calculate_norm_steering(int16_t axis, uint16_t deadzone, bool lin
 
     // apply non-linear transform to the axis
     if (!linear_steering) {
-        return norm_axis * pow(norm_magnitude, 3.0) * max_wheel_value;
+        return (int16_t)(norm_axis * powf(norm_magnitude, 3.0f) * max_wheel_value);
     }
 
-    return norm_axis * norm_magnitude * max_wheel_value;
+    return (int16_t)(norm_axis * norm_magnitude * max_wheel_value);
 }
 
 static void idac_xi_get_analogs(struct idac_io_analog_state *out) {
